🛠 rename: renaming order-related server functions
Ref #53
